feat: correctly support segments
So... apparently from the Unleash instance you get the inlined constraints instead of segments back if you don't send the header. But Edge will always give you the segments. This commit will add the header so that both Edge and Unleash instance will return the same. So we can make sure it always should work with segments.